# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Stratix Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List all documents for the current user
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Register a newly uploaded document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentBody"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateDocument
      tags: [documents]
      summary: Update document metadata (e.g., tags)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDocumentBody"
      responses:
        "200":
          description: Updated document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /documents/search:
    post:
      operationId: searchDocuments
      tags: [documents]
      summary: Semantic search across document chunks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentSearchBody"
      responses:
        "200":
          description: Matching document chunks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentChunk"
  /documents/{id}/process:
    post:
      operationId: processDocument
      tags: [documents]
      summary: Extract text from a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /conversations/{id}/documents:
    get:
      operationId: listConversationDocuments
      tags: [documents]
      summary: List documents linked to a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: linkDocumentToConversation
      tags: [documents]
      summary: Link a document to a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkDocumentBody"
      responses:
        "201":
          description: Linked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    delete:
      operationId: unlinkDocumentFromConversation
      tags: [documents]
      summary: Unlink a document from a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkDocumentBody"
      responses:
        "204":
          description: Unlinked
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Document:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        fileType:
          type: string
        objectKey:
          type: string
        status:
          type: string
        tags:
          type: ['string', 'null']
        chunkCount:
          type: ['integer', 'null']
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - fileType
        - objectKey
        - status
        - createdAt
    CreateDocumentBody:
      type: object
      properties:
        title:
          type: string
        fileType:
          type: string
        objectKey:
          type: string
      required:
        - title
        - fileType
        - objectKey
    UpdateDocumentBody:
      type: object
      properties:
        tags:
          type: string
    DocumentSearchBody:
      type: object
      properties:
        query:
          type: string
        documentIds:
          type: array
          items:
            type: integer
      required:
        - query
    DocumentChunk:
      type: object
      properties:
        id:
          type: integer
        documentId:
          type: integer
        chunkIndex:
          type: integer
        text:
          type: string
        documentTitle:
          type: string
        similarity:
          type: number
      required:
        - id
        - documentId
        - chunkIndex
        - text
        - documentTitle
        - similarity
    LinkDocumentBody:
      type: object
      properties:
        documentId:
          type: integer
      required:
        - documentId
...